Made sdft 16bit

+++ sw/libs/math/inc/sdft.h
@@ -5,7 +5,7 @@
 #include "circular_buffer.h"
 #include <stdint.h>
 
-void sdft(uint8_t new_sample, complex *freq, complex *coeffs);
-uint8_t isdft(complex *freq, complex *coeffs);
+void sdft(uint16_t new_sample, complex *freq, complex *coeffs);
+uint16_t isdft(complex *freq, complex *coeffs);
 
 #endif // SDFT_H
